refactor: moved Const class to Constants folder
