+++ CloneDevOpsTemplate/Controllers/HomeController.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
 using CloneDevOpsTemplate.Models;
-using CloneDevOpsTemplate.Services;
+using CloneDevOpsTemplate.Constants;
 
 namespace CloneDevOpsTemplate.Controllers;
