+++ CloneDevOpsTemplate/Views/Shared/_Header.cshtml
@@ -1,4 +1,4 @@
-@using CloneDevOpsTemplate.Services
+@using CloneDevOpsTemplate.Constants
 @{ 
     var isLoggedin = Context.Session.GetString(Const.SessionKeyAccessToken) != null;
 }
